+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased](https://github.com/qutip/QuantumToolbox.jl/tree/main)
 
 - Support different length for `to` and `from` on GeneralDimensions. ([#448])
+- Extend the `Makie.jl` extension to all the other available backends. ([#450])
 
 ## [v0.30.0]
 Release date: 2025-04-12
@@ -204,3 +205,4 @@ Release date: 2024-11-13
 [#440]: https://github.com/qutip/QuantumToolbox.jl/issues/440
 [#443]: https://github.com/qutip/QuantumToolbox.jl/issues/443
 [#448]: https://github.com/qutip/QuantumToolbox.jl/issues/448
+[#450]: https://github.com/qutip/QuantumToolbox.jl/issues/450